Moving tests over from TestNG to JUnit4.
