+++ integration-tests/src/test/java/integration/util/graylog/GraylogControl.java
@@ -16,23 +16,24 @@
  */
 package integration.util.graylog;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.codec.binary.Base64;
 import org.apache.http.client.utils.URIBuilder;
-import org.testng.SkipException;
 
 import java.io.IOException;
 import java.io.InputStream;
-import java.net.*;
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URISyntaxException;
+import java.net.URL;
 
 public class GraylogControl {
     private URL url;
     private static final int HTTP_TIMEOUT = 1000;
 
     public GraylogControl() {
-        this.url = getServerUrl();
+        this.url = getDefaultServerUrl();
     }
 
     public GraylogControl(URL url) {
@@ -141,31 +142,25 @@ public class GraylogControl {
             InputStream inputStream = connection.getInputStream();
             JsonNode json = mapper.readTree(inputStream);
             return json.get("server_id").toString();
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        } catch (ProtocolException e) {
-            e.printStackTrace();
         } catch (IOException e) {
             e.printStackTrace();
         }
         return "00000000-0000-0000-0000-000000000000";
     }
 
-    public URL getServerUrl() {
-        URL url;
+    public URL getUrl() {
+        return url;
+    }
+
+    private URL getDefaultServerUrl() {
         try {
-            URIBuilder uriBuilder = new URIBuilder(System.getProperty("gl2.baseuri", "http://localhost"));
+            final URIBuilder uriBuilder = new URIBuilder(System.getProperty("gl2.baseuri", "http://localhost"));
             uriBuilder.setPort(Integer.parseInt(System.getProperty("gl2.port", "12900")));
             uriBuilder.setUserInfo(System.getProperty("gl2.admin_user", "admin"), System.getProperty("gl2.admin_password", "admin"));
-            url = uriBuilder.build().toURL();
-        } catch (URISyntaxException e) {
-            throw new SkipException("Invalid URI given. Skipping integration tests.");
-        } catch (MalformedURLException e) {
-            throw new SkipException("Invalid URI given. Skipping integration tests.");
+            return uriBuilder.build().toURL();
+        } catch (URISyntaxException | MalformedURLException e) {
+            throw new RuntimeException("Invalid URI given. Skipping integration tests.");
         }
-        return url;
     }
 
     private static void sleep(long millis) {
